Add plugin support for stats and for initialization.
This commit further fleshes out the pattern for how plugin management
will be done.  There's a new class, PluginCallbackManager, that holds
references to callbacks and provides methods that allow you to call them.

This commit also lists a set of proposed callback points.  The plugin
points are identified by the functions in plugin_registration.py.  We
haven't identified the arguments for most of the callback functions yet,
they are included as placeholders for future work and as discussion
points for the development team.