Overhaul to use SimulationState.
Replaced DataManager's active_ruc, pending_ruc, and so on with a
SimulationState instance.  The state is updated as RUCs and sceds are
executed.

Modified ModelingEngine methods to take a state object instead of ruc
and sced instances with offsets.  Updated implementation in egret_plugin.

Moved reporting of rucs and sceds into a new egret.reporting module.