Report uplift payments.
Payments are reported in thermal_details.csv, in the final timestep of
each day.

Internally, stats are managed slightly differently. At the end of each
timestep, if the completing timestep is the end of an hour or the end
of a day, the stats for all completing period are rolled up first. Stats
are only published after all relevant stats have had a chance to update.
This gives daily stats a chance to update hourly or sced stats.

This is, in fact, how we manage uplift payments. They are recorded in the
sced stats. The value is zero except for the final sced of each day. We
do it this way because we only report per-generator data in thermal_details.csv,
which is written to once per sced.