Split tests into scheduled and triggered, to avoid disabling PR tests.
Github disables workflows with a cron schedule if the repo goes too long
without any changes. It appears to disable all triggers, not just the
schedule trigger. As a result, PRs never get tested and sit there in limbo
forever. This is a) a test to see if that is really what is causing our
PR tests to never run, and b) a fix for the issue.

The fix is to split testing into two separate actions, one for scheduled
tests and one for all other tests. Both call the run-tests workflow.